splitting the file and adding enum also 2 txt files
